requirements.txt
langchain_google_genai
langchain_core
python-dotenv
streamlit
